Initial template
This is a minimal setup for build the future gzdec plugin for both
gstreamer-1.0 and gstreamer-0.10.

The "gzdec_*/" folders contain a slightly modify templates that have been
created using the "gst-project-maker" tool from gst-plugins-bad.

They are two simple makefiles inside gst_{latest,0.10} folders which helps
to download and locally install the gstreamer repository at the desired
version. After that, these repositories could be used to compile and test
the plugins without using the host libraries.

For make easier the build, a top Makefile has been created with targets
that automate the build using host libraries and downloaded libraries.